+++ v3_00/settings.py
@@ -36,6 +36,7 @@ BROWNPALE =  (215, 195, 163)
 BLUEMIDNIGHT = (0, 51, 102)
 GOOGLEMAPSBLUE = (187,197,233)
 DARKRED = (139,0,0)
+PALEGREEN = (177, 237, 171)
 
 # -- New Test Vars For Colour Shades -- 
 # ideally write a function to do this dynamically